refactor: #199 - common 이슈 수정
