+++ src/components/Header/hooks/useSelect.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { SEARCH_KEYS } from '@constants/index';
+import { SEARCH_KEYS } from '@constants';
 
 interface UseSortProps {
   SORT_VALUE?: string;
