+++ src/components/Header/index.tsx
@@ -1,6 +1,6 @@
+import { Icon } from '@components';
 import { PARAM_VALUES, SEARCH_VALUES } from '@constants';
 import styled from '@emotion/styled';
-import Icon from '@components/Icon';
 import { ANGOLA_STYLES } from '@styles/commonStyles';
 import { TabBarList } from './components';
 import { SELECT_OPTION } from './constants';
