+++ src/components/Main/index.tsx
@@ -1,11 +1,10 @@
 import React, { useEffect } from 'react';
 import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
+import { Header, LevelViewer } from '@components';
 import styled from '@emotion/styled';
 import { useCurrentPage, useScrollToTop } from '@hooks';
 import { redirects, routes } from '@routes';
 import { useRecoilValue } from 'recoil';
-import Header from '@components/Header';
-import LevelViewer from '@components/LevelViewer';
 import { useFetchUserArchives } from '@apis/level';
 import { authInfoState } from '@store/auth';
 import { ANGOLA_STYLES } from '@styles/commonStyles';
