--- src/components/NavBar/MenuBar/index.ts
@@ -1,3 +1,0 @@
-export { default as CommonMenuBar } from '@components/NavBar/MenuBar/CommonMenuBar';
-export { default as AuthMenuBar } from '@components/NavBar/MenuBar/AuthMenuBar';
-export { default as NonAuthMenuBar } from '@components/NavBar/MenuBar/NonAuthMenuBar';
