+++ src/components/NavBar/components/CommonMenuBar.tsx
@@ -1,7 +1,5 @@
+import { Icon, LinkButton, SearchBar } from '@components';
 import { useCurrentPage } from '@hooks';
-import Icon from '@components/Icon';
-import LinkButton from '@components/LinkButton';
-import SearchBar from '@components/SearchBar';
 
 const SEARCH = 'search';
 
@@ -23,7 +21,7 @@ const CommonMenuBar = () => {
         <LinkButton to="/search">
           <Icon
             name="search"
-            size={'30'} // TODO:MinwooP - 아이콘 크기도 commonStyle에 유틸화 ?
+            size={'30'}
           />
         </LinkButton>
       )}
