+++ src/components/NavBar/components/index.ts
@@ -1,0 +1,3 @@
+export { default as CommonMenuBar } from '@components/NavBar/components/CommonMenuBar';
+export { default as AuthMenuBar } from '@components/NavBar/components/AuthMenuBar';
+export { default as NonAuthMenuBar } from '@components/NavBar/components/NonAuthMenuBar';
