+++ src/components/NotificationViewer/index.tsx
@@ -1,7 +1,6 @@
 import { MouseEvent } from 'react';
+import { Icon, Spinner } from '@components';
 import styled from '@emotion/styled';
-import Icon from '@components/Icon';
-import Spinner from '@components/Spinner';
 import { ANGOLA_STYLES } from '@styles/commonStyles';
 import DecideNotificationType from './components/DecideNotificationType';
 import { TEXT } from './constants';
