+++ src/components/PostViewer/components/ButtonGroup.tsx
@@ -1,5 +1,5 @@
+import { Icon } from '@components';
 import styled from '@emotion/styled';
-import Icon from '@components/Icon';
 import { useActionInfo, usePostLike } from '@components/PostViewer/hooks';
 import { ANGOLA_STYLES } from '@styles/commonStyles';
 import { BUTTON_GROUP } from '../constants';
