+++ src/pages/HomePage/hooks/useInfiniteScroll.ts
@@ -1,6 +1,6 @@
 import { MutableRefObject, useEffect, useState } from 'react';
+import { Post } from '@type';
 import { useFetchPartPosts } from '@apis/post';
-import { Post } from '@type/index';
 import { INTERACTION_OPTION, LIMIT } from '../constants';
 
 interface useInfinityScroll {
