+++ src/pages/HomePage/index.tsx
@@ -1,12 +1,11 @@
 import { useRef } from 'react';
+import { PostViewer, Spinner } from '@components';
 import styled from '@emotion/styled';
+import { Comment } from '@type';
 import { useRecoilValue } from 'recoil';
-import PostViewer from '@components/PostViewer';
-import Spinner from '@components/Spinner';
 import { authInfoState } from '@store/auth';
 import { calculateLevel } from '@utils/calculateUserLevel';
 import { splitCommentBySeparator } from '@utils/parseDataBySeparator';
-import { Comment } from '@type/index';
 import { useInfiniteScroll } from './hooks';
 
 const HomePage = () => {
