+++ src/pages/MyPage/index.tsx
@@ -1,16 +1,15 @@
 import { useEffect } from 'react';
+import { PostListItem, Spinner } from '@components';
+import { USER_POSTS_TITLE } from '@constants';
 import styled from '@emotion/styled';
 import { useCurrentPage } from '@hooks';
 import { calculateLevel, getUserLevelInfo } from '@utils';
 import { useRecoilValue } from 'recoil';
-import PostListItem from '@components/PostListItem';
-import Spinner from '@components/Spinner';
 import { useFetchDeletePost } from '@apis/post';
 import { useFetchUser } from '@apis/user';
 import { authInfoState } from '@store/auth';
 import { ANGOLA_STYLES } from '@styles/commonStyles';
-import { USER_POSTS_TITLE } from '@constants/index';
-import MyInfo from './MyInfo';
+import MyInfo from './components/MyInfo';
 import { MY_PAGE } from './constants';
 
 const MyPage = () => {
