+++ src/pages/SignUpPage/index.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import { Navigate } from 'react-router-dom';
 import { useNavigate } from 'react-router-dom';
+import { Button } from '@components';
 import styled from '@emotion/styled';
 import { useRecoilValue } from 'recoil';
-import Button from '@components/Button';
 import { authInfoState } from '@store/auth';
 import { ANGOLA_STYLES } from '@styles/commonStyles';
 import {
