--- src/styles/icons/SearchIcon.tsx
@@ -1,28 +1,0 @@
-const SearchIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="42"
-      height="42"
-      viewBox="0 0 42 42"
-      fill="none">
-      <g clipPath="url(#clip0_686_687)">
-        <path
-          d="M41.4258 36.3152L33.2473 28.1367C32.8781 27.7676 32.3777 27.5625 31.8527 27.5625H30.5156C32.7797 24.6668 34.125 21.0246 34.125 17.0625C34.125 7.63711 26.4879 0 17.0625 0C7.63711 0 0 7.63711 0 17.0625C0 26.4879 7.63711 34.125 17.0625 34.125C21.0246 34.125 24.6668 32.7797 27.5625 30.5156V31.8527C27.5625 32.3777 27.7676 32.8781 28.1367 33.2473L36.3152 41.4258C37.0863 42.1969 38.3332 42.1969 39.0961 41.4258L41.4176 39.1043C42.1887 38.3332 42.1887 37.0863 41.4258 36.3152ZM17.0625 27.5625C11.2629 27.5625 6.5625 22.8703 6.5625 17.0625C6.5625 11.2629 11.2547 6.5625 17.0625 6.5625C22.8621 6.5625 27.5625 11.2547 27.5625 17.0625C27.5625 22.8621 22.8703 27.5625 17.0625 27.5625Z"
-          fill="#404040"
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_686_687">
-          <rect
-            width="42"
-            height="42"
-            fill="white"
-          />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-export default SearchIcon;
